backend/requirements.txt
django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
django-environ==0.11.2